+++ package.json
@@ -27,7 +27,7 @@
     "@expo/vector-icons": "^14.0.0",
     "@expo/webpack-config": "~19.0.1",
     "@react-native-async-storage/async-storage": "1.21.0",
-    "@react-native-community/slider": "^4.5.0",
+    "@react-native-community/slider": "4.4.2",
     "@react-native-picker/picker": "2.6.1",
     "@react-navigation/bottom-tabs": "^6.5.12",
     "@react-navigation/material-top-tabs": "^6.6.6",
@@ -35,23 +35,25 @@
     "@react-navigation/native-stack": "^6.9.17",
     "axios": "^1.6.7",
     "expo": "^50.0.0",
-    "expo-font": "~11.10.2",
-    "expo-linear-gradient": "~12.7.1",
+    "expo-font": "~11.10.3",
+    "expo-linear-gradient": "~12.7.2",
     "expo-splash-screen": "~0.26.4",
     "expo-status-bar": "~1.11.1",
-    "expo-updates": "~0.24.10",
+    "expo-updates": "~0.24.12",
     "moti": "^0.19.0",
     "nookies": "^2.5.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.73.4",
+    "react-native": "0.73.6",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-gesture-handler": "~2.14.0",
     "react-native-modal": "^13.0.1",
-    "react-native-pager-view": "^6.2.3",
+    "react-native-pager-view": "6.2.3",
     "react-native-reanimated": "~3.6.2",
+    "react-native-reanimated-carousel": "^3.5.1",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
-    "react-native-svg": "^15.1.0",
+    "react-native-svg": "14.1.0",
     "react-native-svg-charts": "^5.4.0",
     "react-native-tab-view": "^3.5.2",
     "react-native-toast-notifications": "^3.4.0",
@@ -81,7 +83,7 @@
     "husky": "^8.0.1",
     "lint-staged": "15.2.0",
     "prettier": "^3.2.4",
-    "typescript": "^4.8.0"
+    "typescript": "^5.3.0"
   },
   "private": true
 }
